+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Source filter was not enabled unless you were a patreon (unintended).
 * Previously hidden option to disable update notification is now visible in module settings. (Most requested feature ever?).
 * Spell Effects: Aura of Life.
-* Feature Effect: Fighting Style: Interception.
+* Feature Effects: Fighting Style: Interception. Defensive Duelist.
 * Some improvements to DDB tag replacement parsing for magic items.
 
 # 2.9.51